pyproject.toml
[tool.poetry]
name = "tokemak-quant-project"
version = "0.1.0"
description = ""
authors = ["lilioo826"]

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^5.31.1"
jupyterlab = "^3.5.0"
pandas = "^1.5.1"
python-dotenv = "^0.21.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
example = "tokemak_quant_project.example:main"
getBlockNumbers = "tokemak_quant_project.getBlockNumbers:main"
getCurveReserve = "tokemak_quant_project.getCurveReserve:main"
getTradeLogsFromCurve = "tokemak_quant_project.getTradeLogsFromCurve:main"
getTradeLogsFromMavrick = "tokemak_quant_project.getTradeLogsFromMavrick:main"
getMavrickReserve = "tokemak_quant_project.getMavrickReserve:main"
getCurveReserveDaily = "tokemak_quant_project.getCurveReserveDaily:main"
getMavrickReserveDaily = "tokemak_quant_project.getMavrickReserveDaily:main"
calculateCurveLpValue = "tokemak_quant_project.calculateCurveLpValue:main"
